NP-44720 Show messages for published results
