+++ src/translations/nbTranslations.json
@@ -483,7 +483,8 @@
   },
   "my_page": {
     "messages": {
-      "go_to_registration": "Gå til registrering",
+      "files_published": "Fil(er) publisert",
+      "files_rejected": "Fil(er) avvist",
       "messages": "Meldinger",
       "no_messages": "Du har ingen meldinger",
       "ticket_types": {
@@ -1025,6 +1026,7 @@
       "files_are_not_published": "Filer publiseres ikke",
       "files_awaits_approval_one": "{{count}} fil og valgt lisens venter på å bli verifisert",
       "files_awaits_approval_other": "{{count}} filer og valgte lisenser venter på å bli verifisert",
+      "files_awaits_approval_unknown": "Venter på verifisering av fil(er)",
       "files_count": "Filer ({{count}})",
       "find_journal_in_channel_register": "Se informasjon om tidsskriftet i Kanalregisteret",
       "find_publisher_in_channel_register": "Se informasjon om utgiveren i Kanalregisteret",
@@ -1475,7 +1477,7 @@
       "DRAFT_FOR_DELETION": "Slettet",
       "NEW": "Ny",
       "PUBLISHED": "Publisert",
-      "PUBLISHED_METADATA": "Publisert metadata"
+      "PUBLISHED_METADATA": "Metadata publisert"
     },
     "support_description": "Når du ber om brukerstøtte vil meldingen gå til en kurator ved din institusjon. Du kan se besvarelser i <0>Mine Meldinger</0>.",
     "year_published": "Publiseringsår"
